Refacotr auth requests
